Fixed failing tests. Optimized code for better testability
